api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pydantic==2.10.3
python-dotenv==1.0.1
fastapi-cache2>=0.2.1
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
